+++ ohsome_quality_analyst/api/request_models.py
@@ -1,0 +1,108 @@
+import json
+from enum import Enum
+
+import geojson
+import pydantic
+from geojson import Feature, FeatureCollection
+from pydantic import BaseModel
+
+from ohsome_quality_analyst.definitions import (
+    get_dataset_names,
+    get_fid_fields,
+    get_indicator_names,
+    get_report_names,
+    get_topic_keys,
+)
+from ohsome_quality_analyst.topics.models import TopicData
+from ohsome_quality_analyst.utils.helper import snake_to_lower_camel
+from ohsome_quality_analyst.utils.validators import validate_geojson
+
+IndicatorEnum = Enum("IndicatorEnum", {name: name for name in get_indicator_names()})
+ReportEnum = Enum("ReportEnum", {name: name for name in get_report_names()})
+TopicEnum = Enum("TopicEnum", {name: name for name in get_topic_keys()})
+DatasetEnum = Enum("DatasetNames", {name: name for name in get_dataset_names()})
+FidFieldEnum = Enum("FidFieldEnum", {name: name for name in get_fid_fields()})
+
+
+class BaseBpolys(BaseModel):
+    bpolys: Feature | FeatureCollection
+
+    @pydantic.validator("bpolys")
+    @classmethod
+    def validate_bpolys(cls, value) -> FeatureCollection | Feature:
+        obj = geojson.loads(json.dumps(value))
+        validate_geojson(obj)  # Check if exceptions are raised
+        return obj
+
+
+class IndicatorRequest(BaseBpolys):
+    topic_key: TopicEnum = pydantic.Field(
+        ...,
+        title="Topic Key",
+        alias="topic",
+        example="building-count",
+    )
+    include_data: bool = False
+
+    class Config:
+        """Pydantic config class."""
+
+        alias_generator = snake_to_lower_camel
+        # Allow population by field name not just by alias name
+        allow_population_by_field_name = True
+        allow_mutation = False
+        extra = "forbid"
+        schema_extra = {
+            "examples": [
+                {
+                    "topic": "building-count",
+                    "bpolys": {
+                        "type": "Feature",
+                        "geometry": {
+                            "type": "Polygon",
+                            "coordinates": [
+                                [
+                                    [8.674092292785645, 49.40427147224242],
+                                    [8.695850372314453, 49.40427147224242],
+                                    [8.695850372314453, 49.415552187316095],
+                                    [8.674092292785645, 49.415552187316095],
+                                    [8.674092292785645, 49.40427147224242],
+                                ]
+                            ],
+                        },
+                    },
+                }
+            ]
+        }
+
+
+class IndicatorDataRequest(BaseBpolys):
+    """Model for the `/indicators/mapping-saturation/data` endpoint.
+
+    The Topic consists of name, description and data.
+    """
+
+    include_data: bool = False
+    topic: TopicData = pydantic.Field(..., title="Topic", alias="topic")
+
+    class Config:
+        """Pydantic config class."""
+
+        alias_generator = snake_to_lower_camel
+        # Allow population by field name not just by alias name
+        allow_population_by_field_name = True
+        allow_mutation = False
+        extra = "forbid"
+
+
+class ReportRequest(BaseBpolys):
+    include_data: bool = False
+
+    class Config:
+        """Pydantic config class."""
+
+        alias_generator = snake_to_lower_camel
+        # Allow population by field name not just by alias name
+        allow_population_by_field_name = True
+        allow_mutation = False
+        extra = "forbid"